+++ samples/Samples.Posting/Program.cs
@@ -9,8 +9,8 @@ using System.Reflection;
 using Microsoft.Extensions.Logging;
 
 using idunno.AtProto;
-using idunno.AtProto.Repo.Models;
 using idunno.AtProto.Repo;
+
 using idunno.Bluesky;
 using idunno.Bluesky.Embed;
 using idunno.Bluesky.RichText;
@@ -38,7 +38,7 @@ namespace Samples.Posting
             ArgumentException.ThrowIfNullOrEmpty(password);
 
             // Uncomment the next line to route all requests through Fiddler Everywhere
-            // proxyUri = new Uri("http://localhost:8866");
+            proxyUri = new Uri("http://localhost:8866");
 
             // Uncomment the next line to route all requests  through Fiddler Classic
             // proxyUri = new Uri("http://localhost:8888");