+++ src/idunno.AtProto/Repo/Models/ApplyWritesDeleteResponse.cs
@@ -1,8 +1,6 @@
 ﻿// Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
-using System.Text.Json.Serialization;
-
 namespace idunno.AtProto.Repo.Models
 {
     /// <summary>
@@ -10,9 +8,5 @@ namespace idunno.AtProto.Repo.Models
     /// </summary>
     internal sealed record ApplyWritesDeleteResponse : ApplyWritesResponseBase
     {
-        [JsonConstructor]
-        internal ApplyWritesDeleteResponse()
-        {
-        }
     }
 }