+++ src/idunno.AtProto/Repo/Models/DeleteRecordResponse.cs
@@ -5,27 +5,7 @@ using System.Text.Json.Serialization;
 
 namespace idunno.AtProto.Repo.Models
 {
-    /// <summary>
-    /// Encapsulates data returned by the DeleteRecord API call.
-    /// </summary>
-    [JsonUnmappedMemberHandling(JsonUnmappedMemberHandling.Skip)]
-    internal record DeleteRecordResponse
+    internal record DeleteRecordResponse (Commit Commit)
     {
-        /// <summary>
-        /// Creates a new instance of <see cref="DeleteRecordResponse"/>
-        /// </summary>
-        /// <param name="commit">The commit information of the deleteRecord operation.</param>
-        [JsonConstructor]
-        internal DeleteRecordResponse(Commit commit)
-        {
-            Commit = commit;
-        }
-
-        /// <summary>
-        /// Gets the commit information of the deleteRecord operation.
-        /// </summary>
-        [JsonInclude]
-        [JsonRequired]
-        public Commit Commit { get; init; }
     }
 }